requirements.txt
pandas>=2.0
pyarrow>=14.0
openpyxl>=3.1
anthropic>=0.40
pyyaml>=6.0
plotly>=5.18
streamlit>=1.30
requests>=2.31
tqdm>=4.66
unidecode>=1.3
